move stacks/bitcoin specific chainhook types
