+++ components/chainhook-sdk/src/chainhooks/bitcoin/mod.rs
@@ -1,15 +1,12 @@
-use super::types::{
-    BitcoinChainhookInstance, BitcoinPredicateType, DescriptorMatchingRule, ExactMatchingRule,
-    HookAction, InputPredicate, MatchingRule, OrdinalOperations, OrdinalsMetaProtocol,
-    OutputPredicate, StacksOperations,
-};
+use super::types::{ChainhookInstance, ExactMatchingRule, HookAction, MatchingRule};
 use crate::utils::Context;
 
 use bitcoincore_rpc_json::bitcoin::{address::Payload, Address};
 use chainhook_types::{
-    BitcoinBlockData, BitcoinChainEvent, BitcoinTransactionData, BlockIdentifier,
+    BitcoinBlockData, BitcoinChainEvent, BitcoinNetwork, BitcoinTransactionData, BlockIdentifier,
     StacksBaseChainOperation, TransactionIdentifier,
 };
+use schemars::JsonSchema;
 
 use hiro_system_kit::slog;
 
@@ -17,15 +14,136 @@ use miniscript::bitcoin::secp256k1::Secp256k1;
 use miniscript::Descriptor;
 
 use reqwest::{Client, Method};
+use serde::{de, Deserialize, Deserializer};
 use serde_json::Value as JsonValue;
 use std::{
-    collections::{BTreeMap, HashMap},
+    collections::{BTreeMap, HashMap, HashSet},
     str::FromStr,
 };
 
 use reqwest::RequestBuilder;
 
 use hex::FromHex;
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+pub struct BitcoinChainhookSpecification {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub blocks: Option<Vec<u64>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub start_block: Option<u64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub end_block: Option<u64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub expire_after_occurrence: Option<u64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub include_proof: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub include_inputs: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub include_outputs: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub include_witness: Option<bool>,
+    #[serde(rename = "if_this")]
+    pub predicate: BitcoinPredicateType,
+    #[serde(rename = "then_that")]
+    pub action: HookAction,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+pub struct BitcoinChainhookSpecificationNetworkMap {
+    pub uuid: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub owner_uuid: Option<String>,
+    pub name: String,
+    pub version: u32,
+    pub networks: BTreeMap<BitcoinNetwork, BitcoinChainhookSpecification>,
+}
+
+impl BitcoinChainhookSpecificationNetworkMap {
+    pub fn into_specification_for_network(
+        mut self,
+        network: &BitcoinNetwork,
+    ) -> Result<BitcoinChainhookInstance, String> {
+        let spec = self
+            .networks
+            .remove(network)
+            .ok_or("Network unknown".to_string())?;
+        Ok(BitcoinChainhookInstance {
+            uuid: self.uuid,
+            owner_uuid: self.owner_uuid,
+            name: self.name,
+            network: network.clone(),
+            version: self.version,
+            start_block: spec.start_block,
+            end_block: spec.end_block,
+            blocks: spec.blocks,
+            expire_after_occurrence: spec.expire_after_occurrence,
+            predicate: spec.predicate,
+            action: spec.action,
+            include_proof: spec.include_proof.unwrap_or(false),
+            include_inputs: spec.include_inputs.unwrap_or(false),
+            include_outputs: spec.include_outputs.unwrap_or(false),
+            include_witness: spec.include_witness.unwrap_or(false),
+            enabled: false,
+            expired_at: None,
+        })
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub struct BitcoinChainhookInstance {
+    pub uuid: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub owner_uuid: Option<String>,
+    pub name: String,
+    pub network: BitcoinNetwork,
+    pub version: u32,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub blocks: Option<Vec<u64>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub start_block: Option<u64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub end_block: Option<u64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub expire_after_occurrence: Option<u64>,
+    pub predicate: BitcoinPredicateType,
+    pub action: HookAction,
+    pub include_proof: bool,
+    pub include_inputs: bool,
+    pub include_outputs: bool,
+    pub include_witness: bool,
+    pub enabled: bool,
+    pub expired_at: Option<u64>,
+}
+
+impl BitcoinChainhookInstance {
+    pub fn key(&self) -> String {
+        ChainhookInstance::bitcoin_key(&self.uuid)
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct BitcoinTransactionFilterPredicate {
+    pub predicate: BitcoinPredicateType,
+}
+
+impl BitcoinTransactionFilterPredicate {
+    pub fn new(predicate: BitcoinPredicateType) -> BitcoinTransactionFilterPredicate {
+        BitcoinTransactionFilterPredicate { predicate }
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case", tag = "scope")]
+pub enum BitcoinPredicateType {
+    Block,
+    Txid(ExactMatchingRule),
+    Inputs(InputPredicate),
+    Outputs(OutputPredicate),
+    StacksProtocol(StacksOperations),
+    OrdinalsProtocol(OrdinalOperations),
+}
 
 pub struct BitcoinTriggerChainhook<'a> {
     pub chainhook: &'a BitcoinChainhookInstance,
@@ -42,6 +160,177 @@ pub struct BitcoinTransactionPayload {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct BitcoinChainhookPayload {
     pub uuid: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum InputPredicate {
+    Txid(TxinPredicate),
+    WitnessScript(MatchingRule),
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum OutputPredicate {
+    OpReturn(MatchingRule),
+    P2pkh(ExactMatchingRule),
+    P2sh(ExactMatchingRule),
+    P2wpkh(ExactMatchingRule),
+    P2wsh(ExactMatchingRule),
+    Descriptor(DescriptorMatchingRule),
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case", tag = "operation")]
+pub enum StacksOperations {
+    StackerRewarded,
+    BlockCommitted,
+    LeaderRegistered,
+    StxTransferred,
+    StxLocked,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, Hash, PartialEq, Eq, JsonSchema)]
+#[serde(rename_all = "kebab-case")]
+pub enum OrdinalsMetaProtocol {
+    All,
+    #[serde(rename = "brc-20")]
+    Brc20,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
+pub struct InscriptionFeedData {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub meta_protocols: Option<HashSet<OrdinalsMetaProtocol>>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
+#[serde(rename_all = "snake_case", tag = "operation")]
+pub enum OrdinalOperations {
+    InscriptionFeed(InscriptionFeedData),
+}
+
+pub fn get_stacks_canonical_magic_bytes(network: &BitcoinNetwork) -> [u8; 2] {
+    match network {
+        BitcoinNetwork::Mainnet => *b"X2",
+        BitcoinNetwork::Testnet => *b"T2",
+        BitcoinNetwork::Regtest => *b"id",
+        BitcoinNetwork::Signet => unreachable!(),
+    }
+}
+
+pub struct PoxConfig {
+    pub genesis_block_height: u64,
+    pub prepare_phase_len: u64,
+    pub reward_phase_len: u64,
+    pub rewarded_addresses_per_block: usize,
+}
+
+impl PoxConfig {
+    pub fn get_pox_cycle_len(&self) -> u64 {
+        self.prepare_phase_len + self.reward_phase_len
+    }
+
+    pub fn get_pox_cycle_id(&self, block_height: u64) -> u64 {
+        (block_height.saturating_sub(self.genesis_block_height)) / self.get_pox_cycle_len()
+    }
+
+    pub fn get_pos_in_pox_cycle(&self, block_height: u64) -> u64 {
+        (block_height.saturating_sub(self.genesis_block_height)) % self.get_pox_cycle_len()
+    }
+
+    pub fn get_burn_address(&self) -> &str {
+        match self.genesis_block_height {
+            666050 => "1111111111111111111114oLvT2",
+            2000000 => "burn-address-regtest",
+            _ => "burn-address",
+        }
+    }
+}
+const POX_CONFIG_MAINNET: PoxConfig = PoxConfig {
+    genesis_block_height: 666050,
+    prepare_phase_len: 100,
+    reward_phase_len: 2100,
+    rewarded_addresses_per_block: 2,
+};
+
+const POX_CONFIG_TESTNET: PoxConfig = PoxConfig {
+    genesis_block_height: 2000000,
+    prepare_phase_len: 50,
+    reward_phase_len: 1050,
+    rewarded_addresses_per_block: 2,
+};
+
+const POX_CONFIG_DEVNET: PoxConfig = PoxConfig {
+    genesis_block_height: 100,
+    prepare_phase_len: 4,
+    reward_phase_len: 10,
+    rewarded_addresses_per_block: 2,
+};
+
+pub fn get_canonical_pox_config(network: &BitcoinNetwork) -> PoxConfig {
+    match network {
+        BitcoinNetwork::Mainnet => POX_CONFIG_MAINNET,
+        BitcoinNetwork::Testnet => POX_CONFIG_TESTNET,
+        BitcoinNetwork::Regtest => POX_CONFIG_DEVNET,
+        BitcoinNetwork::Signet => unreachable!(),
+    }
+}
+
+#[derive(Debug, Clone, PartialEq)]
+#[repr(u8)]
+pub enum StacksOpcodes {
+    BlockCommit = '[' as u8,
+    KeyRegister = '^' as u8,
+    StackStx = 'x' as u8,
+    PreStx = 'p' as u8,
+    TransferStx = '$' as u8,
+}
+
+impl TryFrom<u8> for StacksOpcodes {
+    type Error = ();
+
+    fn try_from(v: u8) -> Result<Self, Self::Error> {
+        match v {
+            x if x == StacksOpcodes::BlockCommit as u8 => Ok(StacksOpcodes::BlockCommit),
+            x if x == StacksOpcodes::KeyRegister as u8 => Ok(StacksOpcodes::KeyRegister),
+            x if x == StacksOpcodes::StackStx as u8 => Ok(StacksOpcodes::StackStx),
+            x if x == StacksOpcodes::PreStx as u8 => Ok(StacksOpcodes::PreStx),
+            x if x == StacksOpcodes::TransferStx as u8 => Ok(StacksOpcodes::TransferStx),
+            _ => Err(()),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct TxinPredicate {
+    pub txid: String,
+    pub vout: u32,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct DescriptorMatchingRule {
+    // expression defines the bitcoin descriptor.
+    pub expression: String,
+    #[serde(default, deserialize_with = "deserialize_descriptor_range")]
+    pub range: Option<[u32; 2]>,
+}
+
+// deserialize_descriptor_range makes sure that the range value is valid.
+fn deserialize_descriptor_range<'de, D>(deserializer: D) -> Result<Option<[u32; 2]>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let range: [u32; 2] = Deserialize::deserialize(deserializer)?;
+    if !(range[0] < range[1]) {
+        Err(de::Error::custom(
+            "First element of 'range' must be lower than the second element",
+        ))
+    } else {
+        Ok(Some(range))
+    }
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
