+++ components/chainhook-sdk/src/chainhooks/types.rs
@@ -1,13 +1,16 @@
-use std::collections::{BTreeMap, HashSet};
-
 use chainhook_types::{BitcoinNetwork, StacksNetwork};
 use reqwest::Url;
 use serde::ser::{SerializeSeq, Serializer};
-use serde::{de, Deserialize, Deserializer, Serialize};
+use serde::{Deserialize, Serialize};
 
 use schemars::JsonSchema;
 
 use crate::utils::MAX_BLOCK_HEIGHTS_ENTRIES;
+
+use crate::chainhooks::bitcoin::BitcoinChainhookInstance;
+use crate::chainhooks::bitcoin::BitcoinChainhookSpecificationNetworkMap;
+use crate::chainhooks::stacks::StacksChainhookInstance;
+use crate::chainhooks::stacks::StacksChainhookSpecificationNetworkMap;
 
 #[derive(Deserialize, Debug, Clone)]
 pub struct ChainhookConfig {
@@ -30,12 +33,12 @@ impl ChainhookConfig {
     ) -> Result<ChainhookInstance, String> {
         let spec = match hook {
             ChainhookSpecificationNetworkMap::Stacks(hook) => {
-                let spec = hook.into_selected_network_specification(networks.1)?;
+                let spec = hook.into_specification_from_network(networks.1)?;
                 self.stacks_chainhooks.push(spec.clone());
                 ChainhookInstance::Stacks(spec)
             }
             ChainhookSpecificationNetworkMap::Bitcoin(hook) => {
-                let spec = hook.into_selected_network_specification(networks.0)?;
+                let spec = hook.into_specification_for_network(networks.0)?;
                 self.bitcoin_chainhooks.push(spec.clone());
                 ChainhookInstance::Bitcoin(spec)
             }
@@ -193,38 +196,6 @@ impl ChainhookInstance {
     }
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct BitcoinChainhookInstance {
-    pub uuid: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub owner_uuid: Option<String>,
-    pub name: String,
-    pub network: BitcoinNetwork,
-    pub version: u32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub blocks: Option<Vec<u64>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub start_block: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub end_block: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub expire_after_occurrence: Option<u64>,
-    pub predicate: BitcoinPredicateType,
-    pub action: HookAction,
-    pub include_proof: bool,
-    pub include_inputs: bool,
-    pub include_outputs: bool,
-    pub include_witness: bool,
-    pub enabled: bool,
-    pub expired_at: Option<u64>,
-}
-
-impl BitcoinChainhookInstance {
-    pub fn key(&self) -> String {
-        ChainhookInstance::bitcoin_key(&self.uuid)
-    }
-}
-
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case", tag = "chain")]
 pub enum ChainhookSpecificationNetworkMap {
@@ -291,133 +262,6 @@ impl ChainhookSpecificationNetworkMap {
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-pub struct BitcoinChainhookSpecificationNetworkMap {
-    pub uuid: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub owner_uuid: Option<String>,
-    pub name: String,
-    pub version: u32,
-    pub networks: BTreeMap<BitcoinNetwork, BitcoinChainhookSpecification>,
-}
-
-impl BitcoinChainhookSpecificationNetworkMap {
-    pub fn into_selected_network_specification(
-        mut self,
-        network: &BitcoinNetwork,
-    ) -> Result<BitcoinChainhookInstance, String> {
-        let spec = self
-            .networks
-            .remove(network)
-            .ok_or("Network unknown".to_string())?;
-        Ok(BitcoinChainhookInstance {
-            uuid: self.uuid,
-            owner_uuid: self.owner_uuid,
-            name: self.name,
-            network: network.clone(),
-            version: self.version,
-            start_block: spec.start_block,
-            end_block: spec.end_block,
-            blocks: spec.blocks,
-            expire_after_occurrence: spec.expire_after_occurrence,
-            predicate: spec.predicate,
-            action: spec.action,
-            include_proof: spec.include_proof.unwrap_or(false),
-            include_inputs: spec.include_inputs.unwrap_or(false),
-            include_outputs: spec.include_outputs.unwrap_or(false),
-            include_witness: spec.include_witness.unwrap_or(false),
-            enabled: false,
-            expired_at: None,
-        })
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-pub struct BitcoinChainhookSpecification {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub blocks: Option<Vec<u64>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub start_block: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub end_block: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub expire_after_occurrence: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub include_proof: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub include_inputs: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub include_outputs: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub include_witness: Option<bool>,
-    #[serde(rename = "if_this")]
-    pub predicate: BitcoinPredicateType,
-    #[serde(rename = "then_that")]
-    pub action: HookAction,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-pub struct StacksChainhookSpecificationNetworkMap {
-    pub uuid: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub owner_uuid: Option<String>,
-    pub name: String,
-    pub version: u32,
-    pub networks: BTreeMap<StacksNetwork, StacksChainhookSpecification>,
-}
-
-impl StacksChainhookSpecificationNetworkMap {
-    pub fn into_selected_network_specification(
-        mut self,
-        network: &StacksNetwork,
-    ) -> Result<StacksChainhookInstance, String> {
-        let spec = self
-            .networks
-            .remove(network)
-            .ok_or("Network unknown".to_string())?;
-        Ok(StacksChainhookInstance {
-            uuid: self.uuid,
-            owner_uuid: self.owner_uuid,
-            name: self.name,
-            network: network.clone(),
-            version: self.version,
-            start_block: spec.start_block,
-            end_block: spec.end_block,
-            blocks: spec.blocks,
-            capture_all_events: spec.capture_all_events,
-            decode_clarity_values: spec.decode_clarity_values,
-            expire_after_occurrence: spec.expire_after_occurrence,
-            include_contract_abi: spec.include_contract_abi,
-            predicate: spec.predicate,
-            action: spec.action,
-            enabled: false,
-            expired_at: None,
-        })
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-pub struct StacksChainhookSpecification {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub blocks: Option<Vec<u64>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub start_block: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub end_block: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub expire_after_occurrence: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub capture_all_events: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub decode_clarity_values: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub include_contract_abi: Option<bool>,
-    #[serde(rename = "if_this")]
-    pub predicate: StacksPredicate,
-    #[serde(rename = "then_that")]
-    pub action: HookAction,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum HookAction {
     HttpPost(HttpHook),
@@ -451,7 +295,7 @@ pub struct HttpHook {
 pub struct FileHook {
     pub path: String,
 }
-
+// todo: can we remove this struct?
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
 pub struct ScriptTemplate {
     pub instructions: Vec<ScriptInstruction>,
@@ -498,177 +342,6 @@ impl ScriptTemplate {
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-pub struct BitcoinTransactionFilterPredicate {
-    pub predicate: BitcoinPredicateType,
-}
-
-impl BitcoinTransactionFilterPredicate {
-    pub fn new(predicate: BitcoinPredicateType) -> BitcoinTransactionFilterPredicate {
-        BitcoinTransactionFilterPredicate { predicate }
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case", tag = "scope")]
-pub enum BitcoinPredicateType {
-    Block,
-    Txid(ExactMatchingRule),
-    Inputs(InputPredicate),
-    Outputs(OutputPredicate),
-    StacksProtocol(StacksOperations),
-    OrdinalsProtocol(OrdinalOperations),
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum InputPredicate {
-    Txid(TxinPredicate),
-    WitnessScript(MatchingRule),
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum OutputPredicate {
-    OpReturn(MatchingRule),
-    P2pkh(ExactMatchingRule),
-    P2sh(ExactMatchingRule),
-    P2wpkh(ExactMatchingRule),
-    P2wsh(ExactMatchingRule),
-    Descriptor(DescriptorMatchingRule),
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case", tag = "operation")]
-pub enum StacksOperations {
-    StackerRewarded,
-    BlockCommitted,
-    LeaderRegistered,
-    StxTransferred,
-    StxLocked,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, Hash, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "kebab-case")]
-pub enum OrdinalsMetaProtocol {
-    All,
-    #[serde(rename = "brc-20")]
-    Brc20,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
-pub struct InscriptionFeedData {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub meta_protocols: Option<HashSet<OrdinalsMetaProtocol>>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
-#[serde(rename_all = "snake_case", tag = "operation")]
-pub enum OrdinalOperations {
-    InscriptionFeed(InscriptionFeedData),
-}
-
-pub fn get_stacks_canonical_magic_bytes(network: &BitcoinNetwork) -> [u8; 2] {
-    match network {
-        BitcoinNetwork::Mainnet => *b"X2",
-        BitcoinNetwork::Testnet => *b"T2",
-        BitcoinNetwork::Regtest => *b"id",
-        BitcoinNetwork::Signet => unreachable!(),
-    }
-}
-
-pub struct PoxConfig {
-    pub genesis_block_height: u64,
-    pub prepare_phase_len: u64,
-    pub reward_phase_len: u64,
-    pub rewarded_addresses_per_block: usize,
-}
-
-impl PoxConfig {
-    pub fn get_pox_cycle_len(&self) -> u64 {
-        self.prepare_phase_len + self.reward_phase_len
-    }
-
-    pub fn get_pox_cycle_id(&self, block_height: u64) -> u64 {
-        (block_height.saturating_sub(self.genesis_block_height)) / self.get_pox_cycle_len()
-    }
-
-    pub fn get_pos_in_pox_cycle(&self, block_height: u64) -> u64 {
-        (block_height.saturating_sub(self.genesis_block_height)) % self.get_pox_cycle_len()
-    }
-
-    pub fn get_burn_address(&self) -> &str {
-        match self.genesis_block_height {
-            666050 => "1111111111111111111114oLvT2",
-            2000000 => "burn-address-regtest",
-            _ => "burn-address",
-        }
-    }
-}
-
-const POX_CONFIG_MAINNET: PoxConfig = PoxConfig {
-    genesis_block_height: 666050,
-    prepare_phase_len: 100,
-    reward_phase_len: 2100,
-    rewarded_addresses_per_block: 2,
-};
-
-const POX_CONFIG_TESTNET: PoxConfig = PoxConfig {
-    genesis_block_height: 2000000,
-    prepare_phase_len: 50,
-    reward_phase_len: 1050,
-    rewarded_addresses_per_block: 2,
-};
-
-const POX_CONFIG_DEVNET: PoxConfig = PoxConfig {
-    genesis_block_height: 100,
-    prepare_phase_len: 4,
-    reward_phase_len: 10,
-    rewarded_addresses_per_block: 2,
-};
-
-pub fn get_canonical_pox_config(network: &BitcoinNetwork) -> PoxConfig {
-    match network {
-        BitcoinNetwork::Mainnet => POX_CONFIG_MAINNET,
-        BitcoinNetwork::Testnet => POX_CONFIG_TESTNET,
-        BitcoinNetwork::Regtest => POX_CONFIG_DEVNET,
-        BitcoinNetwork::Signet => unreachable!(),
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-#[repr(u8)]
-pub enum StacksOpcodes {
-    BlockCommit = '[' as u8,
-    KeyRegister = '^' as u8,
-    StackStx = 'x' as u8,
-    PreStx = 'p' as u8,
-    TransferStx = '$' as u8,
-}
-
-impl TryFrom<u8> for StacksOpcodes {
-    type Error = ();
-
-    fn try_from(v: u8) -> Result<Self, Self::Error> {
-        match v {
-            x if x == StacksOpcodes::BlockCommit as u8 => Ok(StacksOpcodes::BlockCommit),
-            x if x == StacksOpcodes::KeyRegister as u8 => Ok(StacksOpcodes::KeyRegister),
-            x if x == StacksOpcodes::StackStx as u8 => Ok(StacksOpcodes::StackStx),
-            x if x == StacksOpcodes::PreStx as u8 => Ok(StacksOpcodes::PreStx),
-            x if x == StacksOpcodes::TransferStx as u8 => Ok(StacksOpcodes::TransferStx),
-            _ => Err(()),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct TxinPredicate {
-    pub txid: String,
-    pub vout: u32,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
 pub enum BlockIdentifierIndexRule {
     Equals(u64),
     HigherThan(u64),
@@ -699,148 +372,9 @@ pub enum ExactMatchingRule {
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-pub struct DescriptorMatchingRule {
-    // expression defines the bitcoin descriptor.
-    pub expression: String,
-    #[serde(default, deserialize_with = "deserialize_descriptor_range")]
-    pub range: Option<[u32; 2]>,
-}
-
-// deserialize_descriptor_range makes sure that the range value is valid.
-fn deserialize_descriptor_range<'de, D>(deserializer: D) -> Result<Option<[u32; 2]>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let range: [u32; 2] = Deserialize::deserialize(deserializer)?;
-    if !(range[0] < range[1]) {
-        Err(de::Error::custom(
-            "First element of 'range' must be lower than the second element",
-        ))
-    } else {
-        Ok(Some(range))
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
 pub enum BlockIdentifierHashRule {
     Equals(String),
     BuildsOff(String),
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct StacksChainhookInstance {
-    pub uuid: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub owner_uuid: Option<String>,
-    pub name: String,
-    pub network: StacksNetwork,
-    pub version: u32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub blocks: Option<Vec<u64>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub start_block: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub end_block: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub expire_after_occurrence: Option<u64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub capture_all_events: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub decode_clarity_values: Option<bool>,
-    pub include_contract_abi: Option<bool>,
-    #[serde(rename = "predicate")]
-    pub predicate: StacksPredicate,
-    pub action: HookAction,
-    pub enabled: bool,
-    pub expired_at: Option<u64>,
-}
-
-impl StacksChainhookInstance {
-    pub fn key(&self) -> String {
-        ChainhookInstance::stacks_key(&self.uuid)
-    }
-
-    pub fn is_predicate_targeting_block_header(&self) -> bool {
-        match &self.predicate {
-            StacksPredicate::BlockHeight(_)
-            // | &StacksPredicate::BitcoinBlockHeight(_)
-            => true,
-            _ => false,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-#[serde(tag = "scope")]
-pub enum StacksPredicate {
-    BlockHeight(BlockIdentifierIndexRule),
-    ContractDeployment(StacksContractDeploymentPredicate),
-    ContractCall(StacksContractCallBasedPredicate),
-    PrintEvent(StacksPrintEventBasedPredicate),
-    FtEvent(StacksFtEventBasedPredicate),
-    NftEvent(StacksNftEventBasedPredicate),
-    StxEvent(StacksStxEventBasedPredicate),
-    Txid(ExactMatchingRule),
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct StacksContractCallBasedPredicate {
-    pub contract_identifier: String,
-    pub method: String,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-// #[serde(tag = "type", content = "rule")]
-pub enum StacksContractDeploymentPredicate {
-    Deployer(String),
-    ImplementTrait(StacksTrait),
-}
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum StacksTrait {
-    Sip09,
-    Sip10,
-    #[serde(rename = "*")]
-    Any,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-#[serde(untagged)]
-pub enum StacksPrintEventBasedPredicate {
-    Contains {
-        contract_identifier: String,
-        contains: String,
-    },
-    MatchesRegex {
-        contract_identifier: String,
-        #[serde(rename = "matches_regex")]
-        regex: String,
-    },
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct StacksFtEventBasedPredicate {
-    pub asset_identifier: String,
-    pub actions: Vec<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct StacksNftEventBasedPredicate {
-    pub asset_identifier: String,
-    pub actions: Vec<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct StacksStxEventBasedPredicate {
-    pub actions: Vec<String>,
 }
 
 pub fn opcode_to_hex(asm: &str) -> Option<u8> {
