+++ components/chainhook-sdk/src/indexer/bitcoin/mod.rs
@@ -1,9 +1,8 @@
 use std::time::Duration;
 
-use crate::chainhooks::types::{
+use crate::chainhooks::bitcoin::{
     get_canonical_pox_config, get_stacks_canonical_magic_bytes, PoxConfig, StacksOpcodes,
 };
-
 use crate::observer::BitcoinConfig;
 use crate::utils::Context;
 use bitcoincore_rpc::bitcoin::hashes::Hash;
